requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.5
    # via
    #   cohere
    #   datasets
    #   fsspec
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.5.0
    # via pydantic
annoy==1.17.3
    # via -r requirements.in
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
backoff==2.2.1
    # via
    #   cohere
    #   ragas
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
click==8.1.7
    # via nltk
cohere==4.24
    # via -r requirements.in
dataclasses-json==0.5.14
    # via langchain
datasets==2.14.5
    # via
    #   -r requirements.in
    #   ragas
dill==0.3.7
    # via
    #   datasets
    #   multiprocess
fastavro==1.8.2
    # via cohere
filelock==3.12.4
    # via
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.6.0
    # via
    #   datasets
    #   huggingface-hub
greenlet==2.0.2
    # via sqlalchemy
huggingface-hub==0.17.2
    # via
    #   datasets
    #   sentence-transformers
    #   transformers
idna==3.4
    # via
    #   requests
    #   yarl
importlib-metadata==6.8.0
    # via cohere
jinja2==3.1.2
    # via torch
joblib==1.3.2
    # via
    #   nltk
    #   scikit-learn
langchain==0.0.285
    # via -r requirements.in
langsmith==0.0.37
    # via langchain
markupsafe==2.1.3
    # via jinja2
marshmallow==3.20.1
    # via dataclasses-json
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.15
    # via datasets
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.1
    # via torch
nltk==3.8.1
    # via sentence-transformers
numexpr==2.8.6
    # via langchain
numpy==1.25.2
    # via
    #   datasets
    #   langchain
    #   numexpr
    #   pandas
    #   pyarrow
    #   ragas
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   torchvision
    #   transformers
openai==0.28.0
    # via
    #   -r requirements.in
    #   ragas
packaging==23.1
    # via
    #   datasets
    #   huggingface-hub
    #   marshmallow
    #   transformers
pandas==2.1.0
    # via
    #   -r requirements.in
    #   datasets
pillow==10.0.1
    # via torchvision
protobuf==3.20.0
    # via ragas
pyarrow==13.0.0
    # via datasets
pydantic==2.3.0
    # via
    #   langchain
    #   langsmith
pydantic-core==2.6.3
    # via pydantic
python-dateutil==2.8.2
    # via pandas
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   datasets
    #   huggingface-hub
    #   langchain
    #   transformers
ragas==0.0.6
    # via -r requirements.in
regex==2023.8.8
    # via
    #   nltk
    #   transformers
requests==2.31.0
    # via
    #   cohere
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   langchain
    #   langsmith
    #   openai
    #   torchvision
    #   transformers
safetensors==0.3.3
    # via transformers
scikit-learn==1.3.1
    # via sentence-transformers
scipy==1.11.2
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.2.2
    # via ragas
sentencepiece==0.1.99
    # via sentence-transformers
six==1.16.0
    # via python-dateutil
sqlalchemy==2.0.20
    # via langchain
sympy==1.12
    # via torch
tenacity==8.2.3
    # via langchain
threadpoolctl==3.2.0
    # via scikit-learn
tokenizers==0.13.3
    # via transformers
torch==2.0.1
    # via
    #   sentence-transformers
    #   torchvision
torchvision==0.15.2
    # via sentence-transformers
tqdm==4.66.1
    # via
    #   datasets
    #   huggingface-hub
    #   nltk
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.33.2
    # via
    #   ragas
    #   sentence-transformers
typing-extensions==4.7.1
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via pandas
urllib3==2.0.4
    # via
    #   cohere
    #   requests
xxhash==3.3.0
    # via datasets
yarl==1.9.2
    # via aiohttp
zipp==3.16.2
    # via importlib-metadata
